requirements.txt
altgraph==0.17.3
contourpy==1.0.7
customtkinter==5.2.0
cycler==0.11.0
darkdetect==0.8.0
fonttools==4.39.4
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pefile==2023.2.7
Pillow==9.5.0
pycryptodome==3.18.0
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.5
pyparsing==3.0.9
python-dateutil==2.8.2
pywin32-ctypes==0.2.2
six==1.16.0
zipp==3.15.0
